Cache is a parent for Modules, Interfaces, and Resources
Closes: https://github.com/metarhia/impress/issues/1538
