+++ lib/application.js
@@ -5,7 +5,7 @@ const { path, events } = node;
 const { metavm, metawatch } = metarhia;
 
 const { Interfaces } = require('./interfaces.js');
-const { Cache } = require('./cache.js');
+const { Modules } = require('./modules.js');
 const { Resources } = require('./resources.js');
 
 class Error extends global.Error {
@@ -28,8 +28,8 @@ class Application extends events.EventEmitter {
     this.api = new Interfaces('api', this);
     this.static = new Resources('static', this);
     this.resources = new Resources('resources', this);
-    this.lib = new Cache('lib', this);
-    this.domain = new Cache('domain', this);
+    this.lib = new Modules('lib', this);
+    this.domain = new Modules('domain', this);
     this.starts = [];
 
     this.Application = Application;
