+++ lib/interfaces.js
@@ -1,9 +1,11 @@
 'use strict';
 
 const { node, metarhia } = require('./dependencies.js');
-const { fsp, path, events } = node;
+const { fsp, path } = node;
 const { metavm, metautil } = metarhia;
 const { Procedure } = require('./procedure.js');
+
+const { Cache } = require('./cache.js');
 
 const getSignature = (method) => {
   const src = method.toString();
@@ -12,11 +14,9 @@ const getSignature = (method) => {
   return signature.split(',').map((s) => s.trim());
 };
 
-class Interfaces extends events.EventEmitter {
+class Interfaces extends Cache {
   constructor(place, application) {
-    super();
-    this.path = application.absolute(place);
-    this.application = application;
+    super(place, application);
     this.api = {};
     this.signatures = {};
   }
@@ -86,16 +86,6 @@ class Interfaces extends events.EventEmitter {
     methods[name] = proc;
     internalInterface[name] = method;
     this.cacheSignature(iname + '.' + version, name, method);
-  }
-
-  async load(targetPath = this.path) {
-    const files = await fsp.readdir(targetPath, { withFileTypes: true });
-    for (const file of files) {
-      if (file.name.startsWith('.')) continue;
-      const filePath = path.join(targetPath, file.name);
-      if (file.isDirectory()) await this.load(filePath);
-      else await this.change(filePath);
-    }
   }
 }
