+++ lib/resources.js
@@ -1,16 +1,16 @@
 'use strict';
 
 const { node, metarhia } = require('./dependencies.js');
-const { fsp, path, events } = node;
+const { fsp, path } = node;
 const { metautil } = metarhia;
+
+const { Cache } = require('./cache.js');
 
 const win = process.platform === 'win32';
 
-class Resources extends events.EventEmitter {
+class Resources extends Cache {
   constructor(place, application) {
-    super();
-    this.path = application.absolute(place);
-    this.application = application;
+    super(place, application);
     this.files = new Map();
   }
 
@@ -32,16 +32,6 @@ class Resources extends events.EventEmitter {
       this.files.set(key, data);
     } catch (err) {
       if (err.code !== 'ENOENT') throw err;
-    }
-  }
-
-  async load(targetPath = this.path) {
-    const files = await fsp.readdir(targetPath, { withFileTypes: true });
-    for (const file of files) {
-      if (file.name.startsWith('.')) continue;
-      const filePath = path.join(targetPath, file.name);
-      if (file.isDirectory()) await this.load(filePath);
-      else await this.change(filePath);
     }
   }
 }
