+++ test/modules.js
@@ -2,7 +2,7 @@
 
 const path = require('path');
 const metatests = require('metatests');
-const { Cache } = require('../lib/cache.js');
+const { Modules } = require('../lib/modules.js');
 
 const root = process.cwd();
 
@@ -13,8 +13,8 @@ const application = {
   },
 };
 
-metatests.testAsync('lib/cache load', async (test) => {
-  const cache = new Cache('cache', application);
+metatests.testAsync('lib/modules load', async (test) => {
+  const cache = new Modules('cache', application);
   await cache.load();
   test.strictSame(cache.tree.utils.UNITS.length, 9);
   test.end();
